+++ src/TadasBaltrusaitis_OpenFace/lib/local/FaceAnalyser/src/FaceAnalyser.cpp
@@ -836,20 +836,20 @@ void FaceAnalyser::Reset()
 
 	for( size_t i = 0; i < hog_desc_hist.size(); ++i)
 	{
-		this->hog_desc_hist[i] = cv::Mat_<unsigned int>(hog_desc_hist[i].rows, hog_desc_hist[i].cols, (unsigned int)0);
+		this->hog_desc_hist[i] = cv::Mat_<float>((float)hog_desc_hist[i].rows, (float)hog_desc_hist[i].cols, (float)0);
 		this->hog_hist_sum[i] = 0;
 
 
-		this->face_image_hist[i] = cv::Mat_<unsigned int>(face_image_hist[i].rows, face_image_hist[i].cols, (unsigned int)0);
+		this->face_image_hist[i] = cv::Mat_<float>((float)face_image_hist[i].rows, (float)face_image_hist[i].cols, (float)0);
 		this->face_image_hist_sum[i] = 0;
 
 		// 0 callibration predictions
 		this->au_prediction_correction_count[i] = 0;
-		this->au_prediction_correction_histogram[i] = cv::Mat_<unsigned int>(au_prediction_correction_histogram[i].rows, au_prediction_correction_histogram[i].cols, (unsigned int)0);
+		this->au_prediction_correction_histogram[i] = cv::Mat_<float>((float)au_prediction_correction_histogram[i].rows, (float)au_prediction_correction_histogram[i].cols, (float)0);
 	}
 
 	this->geom_descriptor_median.setTo(cv::Scalar(0));
-	this->geom_desc_hist = cv::Mat_<unsigned int>(geom_desc_hist.rows, geom_desc_hist.cols, (unsigned int)0);
+	this->geom_desc_hist = cv::Mat_<float>((float)geom_desc_hist.rows, (float)geom_desc_hist.cols, (float)0);
 	geom_hist_sum = 0;
 
 	// Reset the predictions
@@ -875,7 +875,7 @@ void FaceAnalyser::Reset()
 	frames_tracking_succ = 0;
 }
 
-void FaceAnalyser::UpdateRunningMedian(cv::Mat_<unsigned int>& histogram, int& hist_count, cv::Mat_<double>& median, const cv::Mat_<double>& descriptor, bool update, int num_bins, double min_val, double max_val)
+void FaceAnalyser::UpdateRunningMedian(cv::Mat_<float>& histogram, int& hist_count, cv::Mat_<double>& median, const cv::Mat_<double>& descriptor, bool update, int num_bins, double min_val, double max_val)
 {
 
 	double length = max_val - min_val;
@@ -885,7 +885,7 @@ void FaceAnalyser::UpdateRunningMedian(cv::Mat_<unsigned int>& histogram, int& h
 	// The median update
 	if(histogram.empty())
 	{
-		histogram = cv::Mat_<unsigned int>(descriptor.cols, num_bins, (unsigned int)0);
+		histogram = cv::Mat_<float>((float)descriptor.cols, (float)num_bins, (float)0);
 		median = descriptor.clone();
 	}
 
@@ -901,7 +901,7 @@ void FaceAnalyser::UpdateRunningMedian(cv::Mat_<unsigned int>& histogram, int& h
 		for(int i = 0; i < histogram.rows; ++i)
 		{
 			int index = (int)converted_descriptor.at<double>(i);
-			histogram.at<unsigned int>(i, index)++;
+			histogram.at<float>(i, index)++;
 		}
 
 		// Update the histogram count
@@ -923,7 +923,7 @@ void FaceAnalyser::UpdateRunningMedian(cv::Mat_<unsigned int>& histogram, int& h
 			int cummulative_sum = 0;
 			for(int j = 0; j < histogram.cols; ++j)
 			{
-				cummulative_sum += histogram.at<unsigned int>(i, j);
+				cummulative_sum += histogram.at<float>(i, j);
 				if(cummulative_sum >= cutoff_point)
 				{
 					median.at<double>(i) = min_val + ((double)j) * (length/((double)num_bins)) + (0.5*(length)/ ((double)num_bins));
@@ -935,7 +935,7 @@ void FaceAnalyser::UpdateRunningMedian(cv::Mat_<unsigned int>& histogram, int& h
 }
 
 
-void FaceAnalyser::ExtractMedian(cv::Mat_<unsigned int>& histogram, int hist_count, cv::Mat_<double>& median, int num_bins, double min_val, double max_val)
+void FaceAnalyser::ExtractMedian(cv::Mat_<float>& histogram, int hist_count, cv::Mat_<double>& median, int num_bins, double min_val, double max_val)
 {
 
 	double length = max_val - min_val;
@@ -963,7 +963,7 @@ void FaceAnalyser::ExtractMedian(cv::Mat_<unsigned int>& histogram, int hist_cou
 			int cummulative_sum = 0;
 			for(int j = 0; j < histogram.cols; ++j)
 			{
-				cummulative_sum += histogram.at<unsigned int>(i, j);
+				cummulative_sum += histogram.at<float>(i, j);
 				if(cummulative_sum > cutoff_point)
 				{
 					median.at<double>(i) = min_val + j * (max_val/num_bins) + (0.5*(length)/num_bins);
@@ -1185,7 +1185,7 @@ void FaceAnalyser::ReadAU(std::string au_model_location)
   
 }
 
-void FaceAnalyser::UpdatePredictionTrack(cv::Mat_<unsigned int>& prediction_corr_histogram, int& prediction_correction_count, vector<double>& correction, const vector<pair<string, double>>& predictions, double ratio, int num_bins, double min_val, double max_val, int min_frames)
+void FaceAnalyser::UpdatePredictionTrack(cv::Mat_<float>& prediction_corr_histogram, int& prediction_correction_count, vector<double>& correction, const vector<pair<string, double>>& predictions, double ratio, int num_bins, double min_val, double max_val, int min_frames)
 {
 	double length = max_val - min_val;
 	if(length < 0)
@@ -1196,7 +1196,7 @@ void FaceAnalyser::UpdatePredictionTrack(cv::Mat_<unsigned int>& prediction_corr
 	// The median update
 	if(prediction_corr_histogram.empty())
 	{
-		prediction_corr_histogram = cv::Mat_<unsigned int>(predictions.size(), num_bins, (unsigned int)0);
+		prediction_corr_histogram = cv::Mat_<float>((float)predictions.size(), (float)num_bins, (float)0);
 	}
 	
 	for(int i = 0; i < prediction_corr_histogram.rows; ++i)
@@ -1211,7 +1211,7 @@ void FaceAnalyser::UpdatePredictionTrack(cv::Mat_<unsigned int>& prediction_corr
 		{
 			index = num_bins - 1;
 		}
-		prediction_corr_histogram.at<unsigned int>(i, index)++;
+		prediction_corr_histogram.at<float>(i, index)++;
 	}
 
 	// Update the histogram count
@@ -1228,7 +1228,7 @@ void FaceAnalyser::UpdatePredictionTrack(cv::Mat_<unsigned int>& prediction_corr
 			int cummulative_sum = 0;
 			for(int j = 0; j < prediction_corr_histogram.cols; ++j)
 			{
-				cummulative_sum += prediction_corr_histogram.at<unsigned int>(i, j);
+				cummulative_sum += prediction_corr_histogram.at<float>(i, j);
 				if(cummulative_sum > cutoff_point)
 				{
 					double corr = min_val + j * (length/num_bins);
@@ -1240,7 +1240,7 @@ void FaceAnalyser::UpdatePredictionTrack(cv::Mat_<unsigned int>& prediction_corr
 	}
 }
 
-void FaceAnalyser::GetSampleHist(cv::Mat_<unsigned int>& prediction_corr_histogram, int prediction_correction_count, vector<double>& sample, double ratio, int num_bins, double min_val, double max_val)
+void FaceAnalyser::GetSampleHist(cv::Mat_<float>& prediction_corr_histogram, int prediction_correction_count, vector<double>& sample, double ratio, int num_bins, double min_val, double max_val)
 {
 
 	double length = max_val - min_val;
@@ -1258,7 +1258,7 @@ void FaceAnalyser::GetSampleHist(cv::Mat_<unsigned int>& prediction_corr_histogr
 		int cummulative_sum = 0;
 		for(int j = 0; j < prediction_corr_histogram.cols; ++j)
 		{
-			cummulative_sum += prediction_corr_histogram.at<unsigned int>(i, j);
+			cummulative_sum += prediction_corr_histogram.at<float>(i, j);
 			if(cummulative_sum > cutoff_point)
 			{
 				double corr = min_val + j * (length/num_bins);
